Store cost as integer, not long
Given that the maximum cost of a single item is 100
an integer can easily accomodate the sum of costs on
any row of 15 elements

+++ src/main/java/com/mobiquityinc/packer/domain/CompositeItem.java
@@ -16,8 +16,8 @@ public class CompositeItem {
         return items.stream().map(Item::getWeight).reduce(0d, (x, y) -> x + y);
     }
 
-    public Long combinedCost() {
-        return items.stream().map(Item::getCost).reduce(0l, (x, y) -> x + y);
+    public Integer combinedCost() {
+        return items.stream().map(Item::getCost).reduce(0, (x, y) -> x + y);
     }
 
     public List<Long> indexNumbers() {
